Moved all Tracer code to Tracer.java
