+++ samples/src/main/java/org/corfudb/samples/graph/GraphDB.java
@@ -3,10 +3,6 @@ package org.corfudb.samples.graph;
 import org.corfudb.runtime.CorfuRuntime;
 import org.corfudb.runtime.collections.SMRMap;
 
-import java.io.BufferedWriter;
-import java.io.File;
-import java.io.FileWriter;
-import java.io.IOException;
 import java.util.*;
 
 /**
@@ -17,80 +13,6 @@ public class GraphDB {
     private SMRMap<UUID, Node> vertices;
     CorfuRuntime rt;
 
-    private Tracer t;
-    private boolean isTracing;
-
-    public class Tracer {
-        String cat;
-        String name;
-        int pid;
-        long tid;
-        long ts;
-        String ph;
-        String[] args;
-        File log;
-
-        public Tracer() {
-            cat = "";
-            name = "";
-            pid = -1;
-            tid = -1;
-            ts = -1;
-            ph = "";
-            args = null;
-            log = new File("performanceLog.json");
-            try {
-                if (!log.exists()) {
-                    log.createNewFile();
-                }
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-        }
-
-        public void updateArgs(String c, String n, int p1, long t1, long t2, String p2, String[] a) {
-            cat = c;
-            name = n;
-            pid = p1;
-            tid = t1;
-            ts = t2;
-            ph = p2;
-            args = a;
-            writeToLog();
-        }
-
-        public void writeToLog() {
-            try {
-                FileWriter fw = new FileWriter(log.getAbsoluteFile(), true);
-                BufferedWriter bw = new BufferedWriter(fw);
-
-                bw.write("{");
-                bw.write("\"cat\": " + "\"" + t.cat + "\",");
-                bw.write("\"pid\": " + t.pid + ",");
-                bw.write("\"tid\": " + t.tid + ",");
-                bw.write("\"ts\": " + t.ts + ",");
-                bw.write("\"ph\": " + "\""  + t.ph + "\",");
-                bw.write("\"name\": " + "\""  + t.name + "\",");
-                bw.write("\"args\": " + t.args + "},\n");
-
-                bw.close();
-                fw.close();
-            } catch (IOException e) {
-                e.printStackTrace();
-            }
-
-            System.out.println("Done");
-        }
-    }
-
-    public void startMethodTrace() {
-        isTracing = true;
-    }
-
-    public void endMethodTrace() {
-        isTracing = false;
-    }
-
     public GraphDB(CorfuRuntime runtime) {
         vertices = runtime.getObjectsView()
                 .build()
@@ -98,8 +20,6 @@ public class GraphDB {
                 .setType(SMRMap.class)  // object class backed by this stream
                 .open();                // instantiate the object!
         rt = runtime;
-        t = new Tracer();
-        isTracing = false;
     }
 
     @Override
@@ -171,22 +91,10 @@ public class GraphDB {
 
     /** Returns an iterable of nodes of all vertices adjacent to v. */
     Iterable<UUID> adjacent(UUID v) {
-        if (isTracing) {
-            t.updateArgs("GraphDBTest", "adjacent", 1, Thread.currentThread().getId(),
-                    System.currentTimeMillis(), "B", null);
-        }
-
-        // begin method
         HashSet<UUID> edgeList = vertices.get(v).getEdges();
         ArrayList<UUID> returnVal = new ArrayList<>();
         for (UUID e : edgeList) {
             returnVal.add(e);
-        }
-        // end method
-
-        if (isTracing) {
-            t.updateArgs("GraphDBTest", "adjacent", 1, Thread.currentThread().getId(),
-                    System.currentTimeMillis(), "E", null);
         }
         return returnVal;
     }
@@ -214,46 +122,16 @@ public class GraphDB {
     }
 
     ArrayList<UUID> preDFS(UUID first) {
-        if (isTracing) {
-            t.updateArgs("GraphDBTest", "preDFS", 1, Thread.currentThread().getId(),
-                    System.currentTimeMillis(), "B", null);
-        }
-
-        // begin method
         ArrayList<UUID> returnVal = dfsHelper(first, "pre", new ArrayList<UUID>(), new ArrayList<UUID>());
-        // end method
-
-        if (isTracing) {
-            t.updateArgs("GraphDBTest", "preDFS", 1, Thread.currentThread().getId(),
-                    System.currentTimeMillis(), "E", null);
-        }
         return returnVal;
     }
 
     ArrayList<UUID> postDFS(UUID first) {
-        if (isTracing) {
-            t.updateArgs("GraphDBTest", "postDFS", 1, Thread.currentThread().getId(),
-                    System.currentTimeMillis(), "B", null);
-        }
-
-        // begin method
         ArrayList<UUID> returnVal = dfsHelper(first, "post", new ArrayList<UUID>(), new ArrayList<UUID>());
-        // end method
-
-        if (isTracing) {
-            t.updateArgs("GraphDBTest", "postDFS", 1, Thread.currentThread().getId(),
-                    System.currentTimeMillis(), "E", null);
-        }
         return returnVal;
     }
 
     ArrayList<UUID> BFS(UUID first) {
-        if (isTracing) {
-            t.updateArgs("GraphDBTest", "BFS", 1, Thread.currentThread().getId(),
-                    System.currentTimeMillis(), "B", null);
-        }
-
-        // begin method
         ArrayList<UUID> ordered = new ArrayList<>();
         ArrayList<UUID> fringe = new ArrayList<>();
         fringe.add(first);
@@ -265,12 +143,6 @@ public class GraphDB {
                     fringe.add(neighbor);
                 }
             }
-        }
-        // end method
-
-        if (isTracing) {
-            t.updateArgs("GraphDBTest", "BFS", 1, Thread.currentThread().getId(),
-                    System.currentTimeMillis(), "E", null);
         }
         return ordered;
     }
